+++ projects/ids-enterprise-typings/lib/searchfield/soho-searchfield.d.ts
@@ -16,7 +16,7 @@ interface SohoSearchFieldOptions extends SohoAutoCompleteOptions {
   showAllResults?: boolean;
 
   /** Displays a dropdown containing categories that can be used to filter results. */
-  categories?: Object[];
+  categories?: SohoSearchFieldCategoryType[];
 
   /** If true, creates a multiselectable Categories list. */
   categoryMultiselect?: boolean;
@@ -44,6 +44,9 @@ interface SohoSearchFieldStatic {
   /**  Gets a complete list of categories in jQuery-collection form. */
   getCategories(): any;
 
+  /** Gets the currently selected categories as data */
+  getCategoryData(onlySelected: boolean): SohoSearchFieldCategory[];
+
   /**  Gets a complete list of categories in jQuery-collection form. */
   getSelectedCategories(): any;
 
@@ -57,7 +60,7 @@ interface SohoSearchFieldStatic {
   element: JQuery;
 
   /** If this component resides within a toolbar, this returns `true` */
-  toolbarParent?: boolean | undefined;
+  toolbarParent?: boolean;
 
   /** Destructor. */
   destroy(): void;
@@ -68,6 +71,19 @@ interface SohoSearchFieldStatic {
   /** Updated */
   updated(settings?: SohoSearchFieldOptions): void;
 }
+
+/**
+ * Search field category.
+ */
+interface SohoSearchFieldCategory {
+    checked: boolean;
+    name: string;
+    id?: string | number;
+    value?: string | number;
+}
+
+/** Search field category type */
+type SohoSearchFieldCategoryType = SohoSearchFieldCategory | string;
 
 /**
  * Type safe SearchField event object.